requirements.txt
authlib==1.6.0
Flask==3.1.1
flask_login==0.6.3
Flask-Migrate==4.1.0
flask_moment==1.0.6
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
google-auth==2.49.1
google-api-python-client~=2.181.0
google-auth-oauthlib~=1.2.2
google-genai~=1.68.0
gunicorn==23.0.0
is_safe_url==1.0
Jinja2==3.1.6
librosa==0.10.1
MarkupSafe==2.1.3
#numpy==2.1.3
pymysql==1.1.2
requests==2.32.5
soundfile==0.12.1
SQLAlchemy>=1.4
vosk==0.3.45
WTForms==3.2.1